[Refactor]: added services for wishlist
