+++ src/Context/wishlist-context.js
@@ -1,7 +1,7 @@
 import { createContext, useContext, useState } from "react";
-import axios from "axios";
 import { useAuth } from "./auth-context";
 import { useEffect } from "react";
+import { getWishlist } from "../services";
 
 const wishlistContext = createContext();
 
@@ -9,67 +9,24 @@ const WishlistProvider = ({ children }) => {
   const [wishlist, setWishlist] = useState([]);
   const { user } = useAuth();
 
-  //GET WISHLIST
+  //GET WISHLIST 
   useEffect(() => {
     user.isUserLoggedIn
       ? (async () => {
-          try {
-            const { data } = await axios.get("api/user/wishlist", {
-              headers: {
-                authorization: user.tokenVal,
-              },
-            });
-            setWishlist(data.wishlist);
-          } catch (error) {
-            console.error("wishlist error", error);
-          }
-        })()
+          
+            const  data  = await getWishlist(user,setWishlist);
+            if(data !== undefined) {
+              setWishlist(data)
+            }
+      })()
       : setWishlist([]);
   }, [user]);
-
-  //ADD to the wishlist
-  const addToWishlist = (product) => {
-    wishlist.find((prod) => prod.id === product.id)
-      ? alert("Already in wishlist")
-      : (async () => {
-          try {
-            const { data } = await axios.post(
-              "api/user/wishlist",
-              { product },
-              {
-                headers: {
-                  authorization: user.tokenVal,
-                },
-              }
-            );
-            setWishlist(data.wishlist);
-          } catch (error) {
-            console.error("wishlist error", error);
-          }
-        })();
-  };
-
-  //delete from wishlist
-  const removeFromWishlist = async (id) => {
-    try {
-      const { data } = await axios.delete(`api/user/wishlist/${id}`, {
-        headers: {
-          authorization: user.tokenVal,
-        },
-      });
-      setWishlist(data.wishlist);
-    } catch (error) {
-      console.error("wishlist error", error);
-    }
-  };
 
   return (
     <wishlistContext.Provider
       value={{
         wishlist,
         setWishlist,
-        addToWishlist,
-        removeFromWishlist,
       }}
     >
       {children}
