+++ src/services/index.js
@@ -1,3 +1,6 @@
-export {signOutHandler} from './Auth/signOutHandler'
-export {logInHandler} from './Auth/logInHandler'
-export {signUpHandler} from './Auth/signUpHandler'
+export { signOutHandler } from "./Auth/signOutHandler";
+export { logInHandler } from "./Auth/logInHandler";
+export { signUpHandler } from "./Auth/signUpHandler";
+export { addToWishlist } from "./wishlist/addToWishlist";
+export { removeFromWishlist } from "./wishlist/removeFromWishlist";
+export { getWishlist } from "./wishlist/getWishlist";
